Refactor tests
I used parametrised tests to avoid duplication.
This is not a very good exemple of parametrized tests because the tests mimics the implementation.
As already said, the exercise is just too simple and the next parametrised tests will be of better quality.
I also customised the test name so the output is more readable.
Now, the code looks better and is generalised.
Let's check the generalisation is correct by adding some tests that I expect to be green right away.