Refactor tests
We avoid duplication in our tests.
This time, the refactoring is a good exemple.
We expressed a rule our code must comply to.
When parametrising tests, it's very important to NOT parametrise the input and the output at the same time because if you do so, you'll have only test which will do everything.
This is really bad for comprehension.
By fixing the output ("Fizz") I still express a behaviour but avoid useless duplication.
When speaking about duplication, we are speaking about duplication of concepts/behaviour; not duplication of text.
Even thought, the lines 31 and 46 are duplication of text (mostly), they express different concepts so I won't try to refactor that.
Now, we have generalised our code. Let's check the generalisation is good by adding new tests that we expect to be green.