Returns "FizzBuzz" on 30
All our tests are green.
Time to refactor.
I could argue that there is three duplication of Fizz / Buzz / 3 and 5 here; and I would be right ! And I could refactor, and use a StringBuilder as you may expect !
But let's say it's not obvious to me for the sake of the exercice.
So, only two duplication of the FizzBuzz concept means I need a third one right ?
Let's add a test for that !