Refactor tests
Well, we can !
It needs a bit of maven configuration but it works !
Now the documentation reads even better !
By reading the tests, we can see one more duplication in the tests.
15 is a multiple of three ! And "FizzBuzz" contains "Fizz". So whatever the multiple of 3, the output should contains "Fizz".
Same goes for 5 and "Buzz".
So, we have some precondition that implies something true about our code. That's a property !
Note that we used the word "whatever". So we can use random input right ?
What about writing a test about that ?