Add a test for any multiple of 3
Great ! We have a test that works with any multiple of three !
We used an assumption in order to validate our preconditions before calling our code.
If the assumption is false, then the code is just ignored. You may check that by running many times the tests. You'll see that sometimes (actually most of the time) the test is ignore.
This is a good reason not to replace your unit tests by property based tests !
An other good reason is that classical unit tests are made of exemple that help to understand what we are talking about.
So, property based tests are great but should NOT replace your existing tests !
"How usefull is it then ?" you may ask. Well... the goal is to replace brain-less tests made by humans like filling a form with random data and see if it fails.
Humans are valuable ! If a machine can do a brain less job, then it should ! A human tester will then have much more time to do exploratory tests.
And by the way, a machine can run hundreds of these tests in a second. That's very effective and the odds of finding a bug (if it exists) will be higher (but not garanteed!).
Let's use our computer power to run that test 100 time.
We could do it by hand, writing a loop, but as soon as an assumption will be proven false, the test will stop so the loop will be near from useless.
Luckily, there is a framework named "JUnitQuickcheck" to the rescue !