Add property on "Buzz"
Ok, we have our two properties !
For the property "Buzz" we have 1 chance over 5 trials to get a correct number. It means that in fact of having 100 tests running, we have only 20 (as a mean).
We can have better odd by biasing our generator. We can do it by using annotations but since our case is very specific, we'll need to create our own generator.
Let's do that !