Created a biased generator for multiples of 5
We created a generator of multiples of 5.
The annotation @MultipleOf5 contrains the use of our generator.
Of course we can do that with the multiple of 3 but that will be left as an exercise to the reader.
If you do that, try to use only une annotation "@MultipleOf" which take a parameter which is the factor.